refactor: extract shared drag bundle image utility
- Move createBundleDragImage to utils/drag_bundle_image.js
- Both event_handlers.js and list_controller.js now use the shared util
- Remove dead .comment-drag-image CSS from comments_popup.css

Addresses review comments on PR #1126.
